+++ webpack.config.js
@@ -30,6 +30,10 @@ const config = {
   resolve: {
     // support reading TypeScript and JavaScript files, 📖 -> https://github.com/TypeStrong/ts-loader
     extensions: ['.ts', '.js'],
+    alias: {
+      canvas: false,
+      'uglify-js': false,
+    },
   },
   module: {
     rules: [